requirements.txt
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
pandas
numpy
django-sslserver-v2
openpyxl
humanize
python-docx
docx2pdf
pytz
redis
celery
django-celery-results
gevent
requests
boto3
psycopg2-binary
django-sslserver-v2
python-dotenv
openai
livekit-agents[silero,turn-detector]~=1.3
livekit-plugins-noise-cancellation~=0.2
livekit-agents[openai]~=1.3
livekit-agents[elevenlabs]~=1.3
livekit-agents[deepgram,cartesia,soniox,gladia,elevenlabs,sarvam,google]~=1.3
twilio